+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.laquerrehugo.app.sim">
+    package="com.laquerrehugo.app.ay">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Material">
         <activity
-            android:name="com.laquerrehugo.app.sim.views.PutActivity"
+            android:name="com.laquerrehugo.app.ay.views.PutActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
