+++ app/src/main/java/com/laquerrehugo/app/ay/services/Contacts.java
@@ -1,11 +1,11 @@
-package com.laquerrehugo.app.sim.services;
+package com.laquerrehugo.app.ay.services;
 
 import android.content.Context;
 import android.content.Intent;
 import android.provider.ContactsContract;
 import android.support.annotation.NonNull;
 
-import com.laquerrehugo.app.sim.models.Contact;
+import com.laquerrehugo.app.ay.models.Contact;
 
 public class Contacts {
     private final Context Context;
