+++ app/src/test/java/com/laquerrehugo/app/ay/InputTypeTest.java
@@ -1,4 +1,4 @@
-package com.laquerrehugo.app.sim;
+package com.laquerrehugo.app.ay;
 
 import org.junit.Test;
